refactor: rename filters.js -> pipes.js
- Renamed nunjucks/src/filters.js to nunjucks/src/pipes.js
- Updated environment.js import from './filters' to './pipes'

All 420 passing
