+++ nunjucks/src/environment.js
@@ -4,7 +4,7 @@ const asap = require('asap');
 const waterfall = require('a-sync-waterfall');
 const lib = require('./lib');
 const compiler = require('./compiler');
-const pipes = require('./filters');
+const pipes = require('./pipes');
 const {FileSystemLoader, WebLoader, PrecompiledLoader} = require('./loaders');
 const tests = require('./tests');
 const globals = require('./globals');
